+++ packages/core/vite.config.ts
@@ -3,5 +3,5 @@ import dts from 'vite-plugin-dts';
 import { baseLibConfig } from '../../vite.config.base.ts';
 
 export default mergeConfig(baseLibConfig, defineConfig({
-  plugins: [dts({ rollupTypes: true })],
+  plugins: [dts({ rollupTypes: true, skipDiagnostics: true })],
 }));